doc comments for everythang + leave some good todo breadcrumbs
