+++ lib/screens/battle_screen/battle_screen_ui.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+/// some cool hacky shit to be able to use built in flutter widgets on top
+///  of your hand drawn Flame canvas
 class BattleScreenUI extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
