+++ lib/screens/battle_screen/noteTrack.dart
@@ -1,0 +1,16 @@
+import 'package:flame/components/component.dart';
+import 'package:flame/components/events/gestures.dart';
+import 'package:flutter/material.dart';
+
+/// The lines that falling notes travel down on
+/// Should render multiple FallingNotes on each track. The FallingNotes should be identical to NoteButton, but without the Tapeable mixin
+class NoteTrack extends PositionComponent {
+  // TODO:
+
+  render(Canvas canvas) {
+  }
+
+  update(double t) {
+  }
+
+}
